# api/types/registry: DecodeAuthConfig: add early returns and improve errors